+++ lib/console.js
@@ -132,12 +132,12 @@ function write(ignoreErrors, stream, string, errorhandler, groupIndent) {
 }
 
 
-// As of v8 5.0.71.32, the combination of rest param, template string
-// and .apply(null, args) benchmarks consistently faster than using
-// the spread operator when calling util.format.
 Console.prototype.log = function log(...args) {
   write(this._ignoreErrors,
         this._stdout,
+        // The performance of .apply and the spread operator seems on par in V8
+        // 6.3 but the spread operator, unlike .apply(), pushes the elements
+        // onto the stack. That is, it makes stack overflows more likely.
         util.format.apply(null, args),
         this._stdoutErrorHandler,
         this[kGroupIndent]);
